requirements.txt
certifi>=2021.10.8
chardet>=3.0.4
click>=8.1.3
Flask>=2.0.3
Flask-HTTPAuth>=4.5.0
gunicorn>=20.0.4
hiredis>=2.0.0
idna>=2.10
itsdangerous>=2.1.1
Jinja2>=3.0.3
MarkupSafe>=2.0.1
netifaces>=0.10.9
paho-mqtt>=1.5.1
pyserial>=3.5
redis>=4.5.5
requests==2.23.0
rmtest>=0.7.0
semantic-version>=2.10.0
uptime>=3.0.1
urllib3>=1.25.11
Werkzeug>=2.0.3
ruamel.yaml>=0.17.21
pyyaml>=6.0
adafruit-circuitpython-ble>=8.3.8
bleak==0.11.0
msgpack>=1.0.3
setuptools==57.0.0
transitions==0.9.0